Ajustes gerais

+++ src/main/java/com/dev/apptite/service/CategoriaService.java
@@ -3,15 +3,15 @@ package com.dev.apptite.service;
 import com.dev.apptite.domain.dto.CategoriaDTO;
 import com.dev.apptite.domain.dto.RestauranteDTO;
 import com.dev.apptite.domain.entity.Categoria;
-import com.dev.apptite.domain.exceptions.BusinessException;
 import com.dev.apptite.domain.exceptions.NotFoundException;
 import com.dev.apptite.domain.mapper.CategoriaMapper;
 import com.dev.apptite.repository.CategoriaRepository;
+import com.dev.apptite.domain.exceptions.BusinessException;
 import lombok.AllArgsConstructor;
+import org.springframework.beans.BeanUtils;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.Optional;
 
 @Service
 @AllArgsConstructor
@@ -22,44 +22,51 @@ public class CategoriaService {
     private final RestauranteService restauranteService;
 
     public CategoriaDTO salvar(CategoriaDTO categoriaDTO) {
-        validacaoCategoria(categoriaDTO);
-        RestauranteDTO restauranteDTO = restauranteService.findById(categoriaDTO.getIdRestaurante());
-        categoriaDTO.setRestaurante(restauranteDTO);
-        Categoria categoria = repository.save(mapper.dtoToEntity(categoriaDTO));
-        return mapper.entityToDTO(categoria);
+
+        associarRestaurante(categoriaDTO);
+        Categoria categoria = mapper.dtoToEntity(categoriaDTO);
+
+        return mapper.entityToDTO(repository.save(categoria));
     }
 
     public List<CategoriaDTO> findAll() {
-        List<Categoria> restaurantes = repository.findAll();
-        return mapper.entityToDTO(restaurantes);
+
+        List<Categoria> categorias = repository.findAll();
+        return mapper.entityToDTO(categorias);
+    }
+
+    public CategoriaDTO update(CategoriaDTO categoriaNovaDTO, Long id) {
+
+        CategoriaDTO categoriaDTO = findById(id);
+
+        BeanUtils.copyProperties(categoriaNovaDTO, categoriaDTO, "idCategoria");
+        Categoria categoria = repository.save(mapper.dtoToEntity(categoriaDTO));
+
+        return mapper.entityToDTO(categoria);
     }
 
     public CategoriaDTO findById(Long id) {
-        Categoria categoria = repository.findById(id).orElseThrow();
+        Categoria categoria = repository.findById(id)
+                .orElseThrow(() -> new NotFoundException("Categoria não encontrada"));
         return mapper.entityToDTO(categoria);
     }
 
     public void delete(Long id) {
+        if (!repository.existsById(id)) {
+            throw new BusinessException("Não é possível deletar. Categoria não encontrada para o ID: " + id);
+        }
         repository.deleteById(id);
     }
 
     public List<CategoriaDTO> findByIdRestaurante(Long idRestaurante) {
-        List<Categoria> categorias = repository.findCategoriasByIdRestaurante(idRestaurante);
+        List<Categoria> categorias = repository.findByIdRestaurante(idRestaurante);
         return mapper.entityToDTO(categorias);
     }
 
-    public List<CategoriaDTO> findAllByIds(List<Long> ids) {
-        List<Categoria> categorias = repository.findByIds(ids);
-        if (categorias.isEmpty()){
-            throw new NotFoundException("base.message.error",List.of("category.not-found.error"));
+    private void associarRestaurante(CategoriaDTO categoriaDTO) {
+        if (categoriaDTO.getIdRestaurante() != null) {
+            RestauranteDTO restauranteDTO = restauranteService.findById(categoriaDTO.getIdRestaurante());
+            categoriaDTO.setRestaurante(restauranteDTO);
         }
-        return mapper.entityToDTO(categorias);
-    }
-
-    private void validacaoCategoria(CategoriaDTO categoriaDTO) {
-        Optional<Categoria> categoriaOptional = repository.findByNome(categoriaDTO.getNome());
-        categoriaOptional.ifPresent(categoria ->{
-            throw new BusinessException("duplicate.category.error");
-        });
     }
 }
